fix use POST_CLEANUP=no for leave resources in cluster

+++ test/e2e/legacy/affinity_toleration.go
@@ -30,6 +30,7 @@ import (
 
 	"github.com/deckhouse/virtualization/api/core/v1alpha2"
 	"github.com/deckhouse/virtualization/api/core/v1alpha2/vmcondition"
+	"github.com/deckhouse/virtualization/test/e2e/internal/config"
 	kc "github.com/deckhouse/virtualization/test/e2e/internal/kubectl"
 	"github.com/deckhouse/virtualization/test/e2e/internal/label"
 )
@@ -71,7 +72,9 @@ var _ = Describe("VirtualMachineAffinityAndToleration", Ordered, label.Legacy(),
 	})
 
 	AfterAll(func() {
-		DeleteTestCaseResources(ns, ResourcesToDelete{KustomizationDir: conf.TestData.AffinityToleration})
+		if config.IsCleanUpNeeded() {
+			DeleteTestCaseResources(ns, ResourcesToDelete{KustomizationDir: conf.TestData.AffinityToleration})
+		}
 	})
 
 	Context("When the virtualization resources are applied:", func() {
