+++ test/e2e/legacy/vm_label_annotation.go
@@ -25,6 +25,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 
 	"github.com/deckhouse/virtualization/api/core/v1alpha2"
+	"github.com/deckhouse/virtualization/test/e2e/internal/config"
 	kc "github.com/deckhouse/virtualization/test/e2e/internal/kubectl"
 	"github.com/deckhouse/virtualization/test/e2e/internal/label"
 )
@@ -57,9 +58,11 @@ var _ = Describe("VirtualMachineLabelAndAnnotation", Ordered, label.Legacy(), fu
 	})
 
 	AfterAll(func() {
-		DeleteTestCaseResources(ns, ResourcesToDelete{
-			KustomizationDir: conf.TestData.VMLabelAnnotation,
-		})
+		if config.IsCleanUpNeeded() {
+			DeleteTestCaseResources(ns, ResourcesToDelete{
+				KustomizationDir: conf.TestData.VMLabelAnnotation,
+			})
+		}
 	})
 
 	Context("When resources are applied", func() {
